requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.2
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.7
django-debug-toolbar==3.2
django-redis-sessions==0.6.2
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.0
idna==2.10
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
py-cpuinfo==7.0.0
pycodestyle==2.7.0
pydantic==1.8.1
pyflakes==2.3.1
pylint==2.7.4
pyparsing==2.4.7
pytest==6.2.2
pytest-benchmark==3.2.3
python-dotenv==0.16.0
pytz==2021.1
redis==3.5.3
regex==2021.3.17
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.23.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
wrapt==1.12.1
